+++ commands_docs/commands_mutable.php
@@ -175,7 +175,7 @@ return array(
 
 			"EXPIREAT key timestamp",
 
-			"Atomically Change the expiration of the <i>key</i> so <i>key</i> expire at specific timestamp.<br />" .
+			"Atomically Change the expiration of the <i>key</i>, so <i>key</i> expire at specific timestamp.<br />" .
 			"If timestamp is in the past, the <i>key</i> is deleted.<br />" .
 			"Note: The command internally GET <i>key</i> first.",
 
@@ -222,41 +222,6 @@ return array(
 			"1.3.4",
 			"READ + WRITE",
 			false,
-			true,
-
-			"mutable"
-	),
-
-	new Cmd(
-			"GETSET",
-
-			"GETSET key value [seconds=0]",
-
-			"Gets the value of the <i>key</i>. Exact match.<br />" .
-			"Then atomically Set <i>key</i> -> <i>value</i> pair.<br />" .
-			"This command is often used to get value of atomic counter and reset its value to zero.",
-			"string",
-			"Value of the key or empty string.",
-			"1.2.11",
-			"READ + WRITE",
-			true,
-			true,
-
-			"mutable"
-	),
-
-	new Cmd(
-			"GETDEL",
-
-			"GETDEL key",
-
-			"Gets the value of the <i>key</i>. Exact match.<br />" .
-			"Then atomically delete the <i>key</i> -> <i>value</i> pair.<br />",
-			"string",
-			"Value of the key or empty string.",
-			"1.2.16",
-			"READ + WRITE",
-			true,
 			true,
 
 			"mutable"